+++ src/zua/mapper/decode.zig
@@ -8,15 +8,15 @@
 const std = @import("std");
 const lua = @import("../../lua/lua.zig");
 
-const Table = @import("../handlers/table.zig");
-const Function = @import("../handlers/function.zig");
-const Userdata = @import("../handlers/userdata.zig").Userdata;
+const Table = @import("../handlers/any/table.zig");
+const Function = @import("../handlers/any/function.zig");
+const Userdata = @import("../handlers/any/userdata.zig").Userdata;
 const Context = @import("../state/context.zig");
 const State = @import("../state/state.zig");
 const Meta = @import("../meta/meta.zig");
 
-const Fn = @import("../typed/fn.zig").Fn;
-const Object = @import("../typed/object.zig").Object;
+const Fn = @import("../handlers/typed/fn.zig").Fn;
+const Object = @import("../handlers/typed/object.zig").Object;
 
 const Mapper = @import("mapper.zig");