+++ src/zua/repl/completion.zig
@@ -7,11 +7,11 @@ const std = @import("std");
 const lua = @import("../../lua/lua.zig");
 const State = @import("../state/state.zig").State;
 const Context = @import("../state/context.zig").Context;
-const Table = @import("../handlers/table.zig").Table;
+const Table = @import("../handlers/any/table.zig").Table;
 const Meta = @import("../meta/meta.zig");
 const Mapper = @import("../mapper/mapper.zig");
 const Native = @import("../functions/native.zig");
-const Object = @import("../typed/object.zig").Object;
+const Object = @import("../handlers/typed/object.zig").Object;
 const isocline = @import("../../isocline/isocline.zig");
 const Config = @import("config.zig");